moved event handling methods to listeners
